fix #563 add confirmation if file already exist

+++ src/gpodder/gtkui/main.py
@@ -1771,6 +1771,14 @@ class gPodder(BuilderWidget, dbus.service.Object):
 
     def save_episodes_as_file(self, episodes):
         def do_save_episode(copy_from, copy_to):
+            if os.path.exists(copy_to):
+                logger.warn(copy_from)
+                logger.warn(copy_to)
+                title = _('File already exist')
+                d = {'filename': os.path.basename(copy_to)}
+                message = _('A file named "%(filename)s" already exist. Do you want to replace it?') % d
+                if not self.show_confirmation(message, title):
+                    return
             try:
                 shutil.copyfile(copy_from, copy_to)
             except (OSError, IOError) as e:
